requirements-build.txt
# Сборка исполняемого файла (Windows)
pyinstaller>=6.0
pillow>=10.0
